Move config to it's own file and prepare for multi-project
